+++ src/accessories/accessory.config.js
@@ -5,7 +5,7 @@ const { validIP, validMAC } = require('../utils/utils');
 const Config = (tvConfig) => {
   const validDisplayCatagories = ['apps', 'channels', 'commands', 'inputs', 'macros'];
   const validSpeakerOutput = ['speaker', 'headphone', 'other'];
-  const validSpeakerAccTypes = ['lightbulb', 'switch'];
+  const validSpeakerAccTypes = ['lightbulb', 'switch', 'fan'];
   const validChannelSources = ['dvbt', 'dvbc', 'dvbs', 'isdbt', 'isdbs', 'isdbc', 'analog'];
   const validInputSources = ['cec', 'component', 'composite', 'hdmi', 'scart', 'widi'];